web/requirements.txt
Django==3.2.10
pipenv==2021.11.23
django-extensions==3.1.5
Werkzeug==2.0.2
pyOpenSSL==21.0.0
mysql-connector-python==8.0.26
pymongo==3.11.0
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0